se integra crear usuarios con JWT
