+++ insert user.sql
@@ -1,3 +1,3 @@
 INSERT INTO public.users(
 	name, password, email, admin, phone, cuit)
-	VALUES ('Ricardo', '$2b$10$kI9jtHzyKNjkWooVKQCptOZTqJAJ9R8edUW7oMknDeIiA9DABsAUm', 'richard@com.com', true, '55698989', 'tres');
+	VALUES ('Ricardo', '$2b$10$guyWP0EFfugpWRqHG2lFFuiHhiGnvtbK.BAWVHpHrsF3tgtBwKrda', 'richard@com.com', true, '55698989', 'tres');
