--- nodemon.json
@@ -1,12 +1,0 @@
-{
-    "restartable": "rs",
-    "ignore": [".git", "node_modules/", "dist/", "coverage/"],
-    "watch": ["src/"],
-    "execMap": {
-      "ts": "node -r ts-node/register"
-    },
-    "env": {
-      "NODE_ENV": "development"
-    },
-    "ext": "js,json,ts"
-  }
