+++ package.json
@@ -4,7 +4,7 @@
   "description": "PF MGF Logistica",
   "main": "app.js",
   "scripts": {
-    "start": "nodemon index.ts",
+    "start": "nodemon index.js",
     "test": "npm test"
   },
   "repository": {
@@ -40,7 +40,6 @@
     "@types/express": "^4.17.13",
     "nodemon": "^2.0.15",
     "ts-node": "^10.4.0",
-    "tslint": "^6.1.3",
-    "typescript": "^4.5.2"
+    "tslint": "^6.1.3"
   }
 }
