--- tsconfig.json
@@ -1,12 +1,0 @@
-{
-    "compilerOptions": {
-      "module": "commonjs",
-      "esModuleInterop": true,
-      "target": "es6",
-      "moduleResolution": "node",
-      "sourceMap": true,
-      "outDir": "dist",
-      "forceConsistentCasingInFileNames": true,
-    },
-    "lib": ["es2015"]
-  }
